Update Files: SessionUserService 추가

+++ src/main/java/com/moneelab/assignment/web/controller/comment/CommentControllerImpl.java
@@ -1,7 +1,10 @@
 package com.moneelab.assignment.web.controller.comment;
 
+import com.moneelab.assignment.config.session.SessionUserService;
 import com.moneelab.assignment.dto.ResponseEntity;
 import com.moneelab.assignment.dto.comment.CommentRequest;
+import com.moneelab.assignment.dto.comment.CommentResponse;
+import com.moneelab.assignment.dto.user.UserResponse;
 import com.moneelab.assignment.service.comment.CommentService;
 
 import javax.servlet.http.HttpServletResponse;
@@ -31,26 +34,39 @@ public class CommentControllerImpl implements CommentController {
      */
     //TODO 사용자인증 구현하면 세션으로 로그인사용자 주입하기
     @Override
-    public ResponseEntity save(CommentRequest commentRequest) {
-        Long savedCommentId = commentService.save(commentRequest, 0L);
+    public ResponseEntity save(CommentRequest commentRequest, SessionUserService sessionService) {
+        UserResponse currentUser = sessionService.getUser();
+
+        Long savedCommentId = commentService.save(commentRequest, currentUser.getUserId());
         return new ResponseEntity(HttpServletResponse.SC_CREATED, savedCommentId);
     }
 
     @Override
-    public ResponseEntity update(Map<String, String> paramMap, CommentRequest commentRequest) {
+    public ResponseEntity update(Map<String, String> paramMap, CommentRequest commentRequest, SessionUserService sessionService) {
         Long commentId = Long.parseLong(paramMap.get("id"));
 
+        validateAuthor(commentId, sessionService);
         commentService.update(commentId, commentRequest);
 
         return new ResponseEntity(HttpServletResponse.SC_NO_CONTENT);
     }
 
     @Override
-    public ResponseEntity delete(Map<String, String> paramMap) {
+    public ResponseEntity delete(Map<String, String> paramMap, SessionUserService sessionService) {
         Long commentId = Long.parseLong(paramMap.get("id"));
 
+        validateAuthor(commentId, sessionService);
         commentService.delete(commentId);
 
         return new ResponseEntity(HttpServletResponse.SC_NO_CONTENT);
+    }
+
+    private void validateAuthor(Long commentId, SessionUserService sessionService) {
+        CommentResponse comment = commentService.findById(commentId);
+
+        UserResponse currentUser = sessionService.getUser();
+        if (!comment.getAuthorId().equals(currentUser.getUserId())) {
+            //TODO 에러
+        }
     }
 }
