+++ src/main/java/com/moneelab/assignment/web/controller/post/PostControllerImpl.java
@@ -1,7 +1,10 @@
 package com.moneelab.assignment.web.controller.post;
 
+import com.moneelab.assignment.config.session.SessionUserService;
 import com.moneelab.assignment.dto.ResponseEntity;
 import com.moneelab.assignment.dto.post.PostRequest;
+import com.moneelab.assignment.dto.post.PostResponse;
+import com.moneelab.assignment.dto.user.UserResponse;
 import com.moneelab.assignment.service.post.PostService;
 
 import javax.servlet.http.HttpServletResponse;
@@ -31,26 +34,39 @@ public class PostControllerImpl implements PostController {
      */
     //TODO 사용자인증 구현하면 세션으로 로그인사용자 주입하기
     @Override
-    public ResponseEntity save(PostRequest postRequest) {
-        Long savedPostId = postService.save(postRequest, 0L);
+    public ResponseEntity save(PostRequest postRequest, SessionUserService sessionService) {
+        UserResponse currentUser = sessionService.getUser();
+
+        Long savedPostId = postService.save(postRequest, currentUser.getUserId());
         return new ResponseEntity(HttpServletResponse.SC_CREATED, savedPostId);
     }
 
     @Override
-    public ResponseEntity update(Map<String, String> paramMap, PostRequest postRequest) {
+    public ResponseEntity update(Map<String, String> paramMap, PostRequest postRequest, SessionUserService sessionService) {
         Long postId = Long.parseLong(paramMap.get("id"));
 
+        validateAuthor(postId, sessionService);
         postService.update(postId, postRequest);
 
         return new ResponseEntity(HttpServletResponse.SC_NO_CONTENT);
     }
 
     @Override
-    public ResponseEntity delete(Map<String, String> paramMap) {
+    public ResponseEntity delete(Map<String, String> paramMap, SessionUserService sessionService) {
         Long postId = Long.parseLong(paramMap.get("id"));
 
+        validateAuthor(postId, sessionService);
         postService.delete(postId);
 
         return new ResponseEntity(HttpServletResponse.SC_NO_CONTENT);
+    }
+
+    private void validateAuthor(Long postId, SessionUserService sessionService) {
+        PostResponse post = postService.findById(postId);
+
+        UserResponse currentUser = sessionService.getUser();
+        if (!post.getAuthorId().equals(currentUser.getUserId())) {
+            //TODO 에러
+        }
     }
 }
